Reference

Nbody2D.Nbody2D

Nbody2D.Box

Nbody2D.Cosmology

Nbody2D.State

Nbody2D.CIC

Nbody2D.Drift

Nbody2D.GRF

Nbody2D.Interp2D

Nbody2D.Kick

Nbody2D.LeapFrog

Nbody2D.Zeldovich

Nbody2D.adot

Nbody2D.growing_mode

Nbody2D.growing_mode

Nbody2D.growing_mode_norm

Nbody2D.Nbody2D — Module

A simple two-dimensional cosmological N-body code.

Nbody2D.Box — Type

The cosmological box, setting the volume of the box L^3 and the number of particles N^3.

Nbody2D.Cosmology — Type

The cosmological background in which the N-body simulation runs with the current Hubble rate H0, the matter density OmegaM, dark energy density OmegaL and the curvature component OmegaK.

Nbody2D.State — Type

The state of the N-body simulation in phase-space.

Nbody2D.CIC — Method

Cloud-in-Cell density estimator.

Nbody2D.Drift — Method

A drift step.

Nbody2D.GRF — Method

Generate an unconstrained Gaussian random field.

Nbody2D.Interp2D — Method

Two dimensional interpolation.

Nbody2D.Kick — Method

A kick step.

Nbody2D.LeapFrog — Method

The Leap Frog integrator.

Nbody2D.Zeldovich — Method

Evaluate the Zel'dovich approximation.

Nbody2D.adot — Method

The derivative of the scale factor.

Nbody2D.growing_mode — Method

The growing mode.

Nbody2D.growing_mode — Method

The growing mode.

Nbody2D.growing_mode_norm — Method

The normalization of the growing mode.