Reference

Nbody2D.Nbody2D

Nbody2D.Box

Nbody2D.CausticSkeleton

Nbody2D.Cosmology

Nbody2D.State

Nbody2D.cGRF

Nbody2D.gridcell

Nbody2D.CIC

Nbody2D.Caustics!

Nbody2D.Drift

Nbody2D.Eulerian

Nbody2D.Eulerian

Nbody2D.Eulerian

Nbody2D.Eulerian

Nbody2D.Eulerian

Nbody2D.GRF

Nbody2D.GRF

Nbody2D.Grad2

Nbody2D.Interp2D

Nbody2D.Kick

Nbody2D.LagrangianGrid

Nbody2D.LeapFrog

Nbody2D.LeapFrogSnapshots

Nbody2D.PolygoniseA!

Nbody2D.PolygoniseA!

Nbody2D.PolygoniseA2!

Nbody2D.PolygoniseA2!

Nbody2D.PolygoniseD!

Nbody2D.PolygoniseD4!

Nbody2D.VertexInterp

Nbody2D.Zeldovich

Nbody2D.adot

Nbody2D.constraintGRF

Nbody2D.convolve

Nbody2D.derivative

Nbody2D.derivative

Nbody2D.derivative

Nbody2D.directional_derivative_1

Nbody2D.directional_derivative_2

Nbody2D.eigenFields123

Nbody2D.first_eigen

Nbody2D.frame

Nbody2D.frame_velocities

Nbody2D.generalizedMoment

Nbody2D.gradient

Nbody2D.growing_mode

Nbody2D.growing_mode_norm

Nbody2D.hessian

Nbody2D.interpolation_cubic

Nbody2D.laplacian

Nbody2D.mean

Nbody2D.meanField

Nbody2D.measureConstraints

Nbody2D.plotMesh

Nbody2D.plotMesh!

Nbody2D.plotSkeleton

Nbody2D.plotSkeleton!

Nbody2D.smooth

Nbody2D.translate

Nbody2D.unwrap_s

Nbody2D.unwrap_x_

Nbody2D.varianceField

Nbody2D.Nbody2D — Module

A simple two-dimensional cosmological N-body code.

Nbody2D.Box — Type

The cosmological box, setting the volume of the box L^2 and the number of particles N^2.

Nbody2D.CausticSkeleton — Type

Caustic skeleton

Nbody2D.Cosmology — Type

The cosmological background in which the N-body simulation runs with the current Hubble rate H0, the matter density OmegaM, dark energy density OmegaL and the curvature component OmegaK.

Nbody2D.State — Type

The state of the N-body simulation in phase-space.

Nbody2D.cGRF — Type

Constrained Gaussian random field object

Nbody2D.gridcell — Type

Gridcell with attributes

Nbody2D.CIC — Method

Cloud-in-Cell density estimator.

Nbody2D.Caustics! — Method

Compute the caustics

Nbody2D.Drift — Method

A drift step.

Nbody2D.Eulerian — Method

Push the skeleton from Lagrangian to Eulerian space

Nbody2D.Eulerian — Method

Push the skeleton from Lagrangian to Eulerian space

Nbody2D.Eulerian — Method

Push the skeleton from Lagrangian to Eulerian space

Nbody2D.Eulerian — Method

Push the skeleton from Lagrangian to Eulerian space

Nbody2D.Eulerian — Method

Push the skeleton from Lagrangian to Eulerian space

Nbody2D.GRF — Method

Generate an unconstrained Gaussian random field.

Nbody2D.GRF — Method

Generate an unconstrained Gaussian random field.

Nbody2D.Grad2 — Method

Finite diference gradient

Nbody2D.Interp2D — Method

Two dimensional interpolation.

Nbody2D.Kick — Method

A kick step.

Nbody2D.LagrangianGrid — Method

Lagrangian grid

Nbody2D.LeapFrog — Method

The Leap Frog integrator.

Nbody2D.LeapFrogSnapshots — Method

The Leap Frog integrator.

Nbody2D.PolygoniseA! — Method

Compute the cusp (A3), swallowtail (A4) and butterfly (A5) caustics

Nbody2D.PolygoniseA! — Method

Compute the cusp (A3), swallowtail (A4) and butterfly (A5) caustics

Nbody2D.PolygoniseA2! — Method

Compute the fold (A2) caustics

Nbody2D.PolygoniseA2! — Method

Compute the fold (A2) caustics

Nbody2D.PolygoniseD! — Method

Compute the umbilic (D4) caustics

Nbody2D.PolygoniseD4! — Method

Compute the umbilic (D4) caustics

Nbody2D.VertexInterp — Method

Interpolate a vertex along a line

Nbody2D.Zeldovich — Method

Evaluate the Zel'dovich approximation.

Nbody2D.adot — Method

The derivative of the scale factor.

Nbody2D.constraintGRF — Method

Generate an unconstrained, constrained Gaussian random field pair.

Nbody2D.convolve — Method

Convolution.

Nbody2D.derivative — Method

Finite difference order dx^4

Nbody2D.derivative — Method

Finite difference order dx^6

Nbody2D.derivative — Method

Finite difference order dx^2

Nbody2D.directional_derivative_1 — Method

Evaluate the first order directional derivative v₁⋅∇λ₁

Nbody2D.directional_derivative_2 — Method

Evaluate the second order directional derivative v₁⋅∇(v₁⋅∇λ₁)

Nbody2D.eigenFields123 — Method

Evaluate first eigenvalue and eigenvector field

Nbody2D.first_eigen — Method

Evaluate first eigenvalue and eigenvector

Nbody2D.frame — Function

frame(coords_q, coords_x, L, pad=0.05)

Cosmological N-body simulations normally work with periodic boundary conditions. This function adds a frame of periodic particle positions around the simulation box to implement periodicity in the Delaunay tesselation. pad specifies the width of the frame in units of the simulation box size L.

Nbody2D.frame_velocities — Function

frame_velocities(coords_x, velocities, L, pad=0.05)

Cosmological N-body simulations normally work with periodic boundary conditions. This function adds a frame of velocities (corresponding to the periodic particle positions) around the simulation box to implement periodicity in the Delaunay tesselation. pad specifies the width of the frame in units of the simulation box size L.

Nbody2D.generalizedMoment — Method

Generalzed moments.

Nbody2D.gradient — Method

Gradient of a scalar field

Nbody2D.growing_mode — Method

The growing mode.

Nbody2D.growing_mode_norm — Method

The normalization of the growing mode.

Nbody2D.hessian — Method

Hessian of a scalar field

Nbody2D.interpolation_cubic — Method

Cubic interpolation of a regular grid

Nbody2D.laplacian — Method

Laplace operator.

Nbody2D.mean — Method

Mean of an Array.

Nbody2D.meanField — Method

The mean field given the constraints c.

Nbody2D.measureConstraints — Method

Measure the constraints.

Nbody2D.plotMesh — Function

Mesh plot of the N-body simulation

Nbody2D.plotMesh! — Function

Mesh plot of the N-body simulation

Nbody2D.plotSkeleton — Function

Plot Caustic Skeleton

Nbody2D.plotSkeleton! — Function

Plot Caustic Skeleton

Nbody2D.smooth — Method

Smooth an Array.

Nbody2D.translate — Method

translate(coords_q, coords_x, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function shifts both the initial and final positions of the particles such that they are located in the simulation box in Eulerian space.

Nbody2D.unwrap_s — Method

unwrap_s(s, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function undoes the wrapping of particles around the box in the displacement field, assuming the particles did not travel more than half the size of the box.

Nbody2D.unwrap_x_ — Method

unwrap_x_(q, x, L)

Cosmological N-body simulations normally work with periodic boundary conditions. This function undoes the wrapping of particles around the box, assuming the particles did not travel more than half the size of the box.

Nbody2D.varianceField — Method

The variance field of the constraints.